+++ __tests__/app/dashboard.test.tsx
@@ -19,6 +19,11 @@ const mockFetchData = {
   totalRolloutExperiments: 0,
 };
 
+// XXX these issues seem to be in part related to asynchrony and
+// have so far turned out be difficult to resolve. Let's see if upgrading to
+// React 19 makes this easier to untangle so we can re-enabled. It's
+// conceivable that there are bugs in the framework software or bundler that
+// are in play.
 describe.skip("Dashboard", () => {
   it("all timeline pill ids exist in the Dashboard component in /", async () => {
     const dashboard = await render(await (<Dashboard {...mockFetchData} />));