+++ __tests__/app/message-table.test.tsx
@@ -14,7 +14,15 @@ import { NimbusRecipe } from "@/lib/nimbusRecipe";
 import { compareSurfacesFn } from "@/lib/messageUtils";
 import userEvent from "@testing-library/user-event";
 
+// These are part of the mock control API
+// eslint-disable-next-line jest/no-mocks-import
+import { resetMockState, setMockPlatform } from "@/lib/__mocks__/sdk";
+
 jest.mock("../../lib/sdk");
+
+afterEach(() => {
+  resetMockState();
+});
 
 describe("MessageTable", () => {
   describe("ExperimentColumns", () => {
@@ -178,6 +186,7 @@ describe("MessageTable", () => {
     });
 
     it("displays CTR percentages when Looker dashboard exists and CTR is defined", async () => {
+      setMockPlatform("firefox-desktop");
       const nimbusRecipeCollection = new NimbusRecipeCollection();
       nimbusRecipeCollection.recipes = [
         new NimbusRecipe(ExperimentFakes.recipe()),
@@ -186,6 +195,7 @@ describe("MessageTable", () => {
         (await nimbusRecipeCollection.getExperimentAndBranchInfos()) as RecipeInfo[];
       // Setting fake dashboard link in order to render in MessageTable
       recipeInfos[0].branches[0].ctrDashboardLink = "test link";
+
       render(<MessageTable columns={experimentColumns} data={recipeInfos} />);
 
       const toggleButton = screen.getByTestId("toggleAllRowsButton");