+++ lib/messageUtils.ts
@@ -1,97 +1,182 @@
 import { FxMSMessageInfo } from "@/app/columns";
 import { getLookerSubmissionTimestampDateFilter } from "./lookerUtils";
+import { Platform } from "./types";
 
 export type SurfaceData = {
   surface: string;
+  platform: Platform;
+  lookerDateFilterPropertyName: string;
+  lookerDashboardId: string;
   tagColor?: string;
   docs?: string;
 };
 
-export function getSurfaceDataForTemplate(template: string): SurfaceData {
+/**
+ * @param surfaceOrTemplate a desktop template or mobile surface
+ * @returns data pertaining to the given desktop template or mobile surface
+ * (ie. surface display name, platform, Looker date filter property name,
+ * Looker dashboard ID, surface tag background colour)
+ */
+export function getSurfaceData(surfaceOrTemplate: string): SurfaceData {
   const surfaceData: Record<string, SurfaceData> = {
     aboutwelcome: {
       surface: "About:Welcome Page (1st screen)",
+      platform: "firefox-desktop",
+      lookerDateFilterPropertyName: "event_counts.submission_timestamp_date",
+      lookerDashboardId: "1818",
       tagColor: "bg-red-400",
     },
     cfr: {
       surface: "Contextual Feature Recommendation",
+      platform: "firefox-desktop",
+      lookerDateFilterPropertyName: "event_counts.submission_timestamp_date",
+      lookerDashboardId: "1818",
       tagColor: "bg-red-200",
     },
     cfr_doorhanger: {
       surface: "Doorhanger",
+      platform: "firefox-desktop",
+      lookerDateFilterPropertyName: "event_counts.submission_timestamp_date",
+      lookerDashboardId: "1818",
       tagColor: "bg-orange-200",
       docs: "https://experimenter.info/messaging/desktop-messaging-surfaces/#doorhanger",
     },
     defaultaboutwelcome: {
       surface: "Default About:Welcome Message (1st screen)",
+      platform: "firefox-desktop",
+      lookerDateFilterPropertyName: "event_counts.submission_timestamp_date",
+      lookerDashboardId: "1818",
       tagColor: "bg-orange-400",
     },
     feature_callout: {
       surface: "Feature Callout (1st screen)",
+      platform: "firefox-desktop",
+      lookerDateFilterPropertyName: "event_counts.submission_timestamp_date",
+      lookerDashboardId: "1818",
       tagColor: "bg-yellow-300",
       docs: "https://experimenter.info/messaging/desktop-messaging-surfaces/#feature-callouts",
     },
     infobar: {
       surface: "InfoBar",
+      platform: "firefox-desktop",
+      lookerDateFilterPropertyName: "messaging_system.submission_date",
+      lookerDashboardId: "2267",
       tagColor: "bg-lime-300",
       docs: "https://experimenter.info/messaging/desktop-messaging-surfaces/#infobar",
     },
     menu: {
       surface: "Menu Messages",
+      platform: "firefox-desktop",
+      lookerDateFilterPropertyName: "event_counts.submission_timestamp_date",
+      lookerDashboardId: "1818",
       tagColor: "bg-pink-300",
     },
     milestone_message: {
       surface: "Milestone Messages",
+      platform: "firefox-desktop",
+      lookerDateFilterPropertyName: "event_counts.submission_timestamp_date",
+      lookerDashboardId: "1818",
       tagColor: "bg-green-400",
     },
     multi: {
       surface: "1st of Multiple Messages",
+      platform: "firefox-desktop",
+      lookerDateFilterPropertyName: "event_counts.submission_timestamp_date",
+      lookerDashboardId: "1818",
       tagColor: "bg-teal-300",
     },
     pb_newtab: {
       surface: "Private Browsing New Tab",
+      platform: "firefox-desktop",
+      lookerDateFilterPropertyName: "event_counts.submission_timestamp_date",
+      lookerDashboardId: "1818",
       tagColor: "bg-sky-400",
       docs: "https://experimenter.info/messaging/desktop-messaging-surfaces/#privatebrowsing",
     },
     protections_panel: {
       surface: "Protections Dropdown Panel",
+      platform: "firefox-desktop",
+      lookerDateFilterPropertyName: "event_counts.submission_timestamp_date",
+      lookerDashboardId: "1818",
       tagColor: "bg-blue-500",
     },
     rtamo: {
       surface: "Return to AMO",
+      platform: "firefox-desktop",
+      lookerDateFilterPropertyName: "event_counts.submission_timestamp_date",
+      lookerDashboardId: "1818",
       tagColor: "bg-violet-200",
     },
     toast_notification: {
       surface: "Toast Notification",
+      platform: "firefox-desktop",
+      lookerDateFilterPropertyName: "event_counts.submission_timestamp_date",
+      lookerDashboardId: "1818",
       tagColor: "bg-indigo-400",
     },
-    toolbar_badge: { surface: "Toolbar Badge", tagColor: "bg-purple-400" },
+    toolbar_badge: {
+      surface: "Toolbar Badge",
+      platform: "firefox-desktop",
+      lookerDateFilterPropertyName: "event_counts.submission_timestamp_date",
+      lookerDashboardId: "1818",
+      tagColor: "bg-purple-400",
+    },
     // XXX Consider removing after we start reading JSON from remote settings
-    "toolbar-badge": { surface: "Toolbar Badge", tagColor: "bg-purple-400" },
+    "toolbar-badge": {
+      surface: "Toolbar Badge",
+      platform: "firefox-desktop",
+      lookerDateFilterPropertyName: "event_counts.submission_timestamp_date",
+      lookerDashboardId: "1818",
+      tagColor: "bg-purple-400",
+    },
     spotlight: {
       surface: "Spotlight Modal Dialog",
+      platform: "firefox-desktop",
+      lookerDateFilterPropertyName: "event_counts.submission_timestamp_date",
+      lookerDashboardId: "1818",
       tagColor: "bg-pink-400",
       docs: "https://experimenter.info/messaging/desktop-messaging-surfaces/#multistage-spotlight",
     },
     survey: {
       surface: "Survey",
+      platform: "fenix",
+      lookerDateFilterPropertyName: "events.submission_date",
+      lookerDashboardId: "2303",
       tagColor: "bg-cyan-400",
     },
     update_action: {
       surface: "Moments Page",
+      platform: "firefox-desktop",
+      lookerDateFilterPropertyName: "event_counts.submission_timestamp_date",
+      lookerDashboardId: "1818",
       tagColor: "bg-rose-400",
       docs: "https://experimenter.info/messaging/desktop-messaging-surfaces/#moments-pages",
     },
-    "whats-new-panel": { surface: "What's New Panel", tagColor: "bg-sky-200" },
-    whatsNewPage: { surface: "What's New Page", tagColor: "bg-fuchsia-300" },
+    "whats-new-panel": {
+      surface: "What's New Panel",
+      platform: "firefox-desktop",
+      lookerDateFilterPropertyName: "event_counts.submission_timestamp_date",
+      lookerDashboardId: "1818",
+      tagColor: "bg-sky-200",
+    },
+    whatsNewPage: {
+      surface: "What's New Page",
+      platform: "firefox-desktop",
+      lookerDateFilterPropertyName: "event_counts.submission_timestamp_date",
+      lookerDashboardId: "1818",
+      tagColor: "bg-fuchsia-300",
+    },
   };
 
-  if (template in surfaceData) {
-    return surfaceData[template];
+  if (surfaceOrTemplate in surfaceData) {
+    return surfaceData[surfaceOrTemplate];
   }
 
   return {
-    surface: template,
+    surface: surfaceOrTemplate,
+    platform: "firefox-desktop",
+    lookerDateFilterPropertyName: "event_counts.submission_timestamp_date",
+    lookerDashboardId: "1818",
   };
 }
 
@@ -117,8 +202,8 @@ export function _isAboutWelcomeTemplate(template: string): boolean {
   return aboutWelcomeSurfaces.includes(template);
 }
 
-export function getAndroidDashboard(
-  template: string,
+export function getAndroidDashboardLink(
+  surface: string,
   msgIdPrefix: string,
   channel?: string,
   experiment?: string,
@@ -135,9 +220,7 @@ export function getAndroidDashboard(
     isCompleted,
   );
 
-  // XXX consider using a similar function like getDashboardIdForTemplate for
-  // android dashboards to get dashboardId
-  const dashboardId = 2303; // messages/push notification
+  const dashboardId = getSurfaceData(surface).lookerDashboardId; // messages/push notification
   let baseUrl = `https://mozilla.cloud.looker.com/dashboards/${dashboardId}`;
   let paramObj;
 
@@ -158,7 +241,7 @@ export function getAndroidDashboard(
   };
 
   // XXX we really handle all messaging surfaces, at least in theory
-  if (template !== "survey") return undefined;
+  if (surface !== "survey") return undefined;
 
   if (paramObj) {
     const params = new URLSearchParams(Object.entries(paramObj));
@@ -170,7 +253,7 @@ export function getAndroidDashboard(
   return undefined;
 }
 
-export function getDashboard(
+export function getDesktopDashboardLink(
   template: string,
   msgId: string,
   channel?: string,
@@ -187,7 +270,7 @@ export function getDashboard(
     endDate,
     isCompleted,
   );
-  const dashboardId = getDashboardIdForTemplate(template);
+  const dashboardId = getDashboardIdForSurface(template);
   let baseUrl = `https://mozilla.cloud.looker.com/dashboards/${dashboardId}`;
   let paramObj;
 
@@ -269,14 +352,10 @@ export function getPreviewLink(message: any): string {
 
 /**
  * XXX consider moving this function inside looker.ts
- * @returns the Looker dashboard ID for a given message template
+ * @returns the Looker dashboard ID for a given desktop template or mobile surface
  */
-export function getDashboardIdForTemplate(template: string) {
-  if (template === "infobar") {
-    return "2267";
-  } else {
-    return "1818";
-  }
+export function getDashboardIdForSurface(surfaceOrTemplate: string) {
+  return getSurfaceData(surfaceOrTemplate).lookerDashboardId;
 }
 
 /**