+++ next.config.js
@@ -34,6 +34,9 @@ const cspHeaderDev = `
 
 const nextConfig = {
   staticPageGenerationTimeout: 240,
+  eslint: {
+    dirs: ["pages", "app", "components", "lib", "src", "__tests__"],
+  },
   async headers() {
     return [
       {